Better url generation. olx.py split into category.py, offer.py, utils.py, __init__.py Added example.py
